requirements.txt
six
djangorestframework>=3.3,<3.8
